fix: api_keys.used_tokens 迁移 BigInt——schema 五方言与 init.sql 同步改为 64 位整数，admin API 序列化补 Number 转换（keys 列表/创建/更新、usage 统计、导出 replacer），worker 认证额度比较与 key-reset 重置适配 bigint，相关测试断言同步

+++ init.sql
@@ -39,7 +39,7 @@ CREATE TABLE IF NOT EXISTS "api_keys" (
     "id" TEXT NOT NULL PRIMARY KEY,
     "key" TEXT NOT NULL,
     "name" TEXT NOT NULL,
-    "used_tokens" INTEGER NOT NULL DEFAULT 0,
+    "used_tokens" BIGINT NOT NULL DEFAULT 0,
     "token_limit" INTEGER,
     "rpm_limit" INTEGER,
     "tpm_limit" INTEGER,
